Add MOTD and prepend '\n' before each shell prompt
Add precmd() hook that will prepend a newline character before every
prompt if the last command entered by the user was not `clear`.

Add a MOTD so that the terminal does not look ugly when there is a
newline character at the top of the terminal when it starts up.